+++ CHANGELOG.md
@@ -16,6 +16,13 @@ This release will contain consensus-breaking changes.
   resolving several bugs in the old parser and improving performance.
 - Documentation will indicate explicitly which Clarity version introduced each
   keyword or function. 
+- Clarity2 improvements to traits (see #3251 for details):
+  * Trait values can be passed to compatible sub-trait types
+  * Traits can be embedded in compound types, e.g. `(optional <my-trait>)
+  * Traits can be assigned to a let-variable
+- Fixes to unexpected behavior in traits
+  * A trait with duplicate function names is now an error (#3214)
+  * Aliased trait names do not interfere with local trait definitions (#3215)
 
 ## Upcoming